requirements.txt
crewai==0.86.0
crewai-tools==0.17.0
langchain==0.3.11
openai==1.57.3
langchain-openai==0.2.12
pydantic==2.10.3
PyYAML==6.0.2
dnspython==2.7.0
python-dotenv==1.0.1
pymongo==4.10.1
google-api-python-client==2.155.0
google-auth==2.37.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
requests==2.32.3
colorama==0.4.6
pytest==8.3.4
black==24.10.0
flake8==7.1.1